+++ src/gui/clipboardbrowserplaceholder.h
@@ -65,6 +65,7 @@ protected:
     bool event(QEvent *event) override;
 
 private:
+    void scheduleBrowserCreation();
     void setActiveWidget(QWidget *widget);
 
     bool canExpire() const;
@@ -80,6 +81,7 @@ private:
     void onFilterProgressChanged(int percent);
 
     ClipboardBrowser *m_browser = nullptr;
+    bool m_browserCreationScheduled = false;
     QPushButton *m_loadButton = nullptr;
     QProgressBar *m_filterProgressBar = nullptr;